Implements a number of additional tests and file manager functionality for touching files.
This commit implements a number of additional tests (and as a result of those fixes) as well as file manager functionality for touching files. Furthermore, an additional path representation property has been added which describes whether or not the original path specified obviously referred to a directory (i.e. it had a terminating forward slash).