Remove useless unit tests and add S3Adapter.GetFileAsync
This commit removes a selection of unit tests that fundamentally tested
the same exact code. Got bored of writing them, and by testing one and
actually using my eyes I'm sure I can see when things are using the
methods that they should be.

In addition, this commit implements the S3Adapter.GetFileAsync method which,
at the minute, doesn't do an awful lot.