Continue implementing directory copies
Continued the implementation of directory copies within the S3 adapter.

Also added a better way to paginate and act on S3 list responses as the original way
would only work with actions that modified the original files (i.e. by deleting
them).

Learned a number of things that have been added to the to-do list:
	* The root path should probably be a first party citizens within tests rather
	  than being occassionally used.
	* Need to investigate whether or not copying/moving of directories to
	  somewhere that already exists should be allowed to happen.
	* Need to standardise where CombineRootAndRequestedPath is called as I'm
	  getting confused looking at the code as to when it is actually happening.
	  Will probably establish a standard that all paths are combined at the
	  top level functions, i.e. CopyFileAsync, CreateDirectoryAsync etc.