Add root path tests, randomise names more, fix tests.
Added tests for root paths, ensuring that root paths are tested for each and every function instead of
randomly here and there. Located a few issues a result which is a good sign.

Replaced GUIDs and fixed names in files and directories along with the root path to be randomised,
alpha-numeric strings of a configurable (or random) size. This should stop any conflicts, ever, between
the tests (if there could possibly be any).

Fixed a number of tests that were broken and also broke as a result of the changes made.