ISSUE-3 - Move path and root path combination to the manager level
Moved the path and root path combination out of the adapter and into the
top level managers.

Implemented a nice way to clone and update paths for requests.

Modified the tests to cater for this new way of doing things.

Moved some namespaces around that weren't in the correct place.

Issue #3