fix(s3-adapter): fix iteration/listing returning more paths than applicable
As S3 directories are virtual and only FILE paths are returned from queries, we have
to manually construct the path tree in order to return a standardised response from
the S3 adapter.

Where a given list request was X levels deep, we were returning all directories from
the root directory -> X as responses when we shouldn't have been.